Added authorization to CustomerInfo controllers

+++ RentalDataWarehouse/Controllers/CustomerInfoControllers/CustomerInfoController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using RentalDataWarehouse.Data;
@@ -13,15 +14,21 @@ namespace RentalDataWarehouse.Controllers
     public class CustomerInfoController : Controller
     {
         private readonly ApplicationDbContext _context;
+        private readonly UserManager<ApplicationUser> _userManager;
 
-        public CustomerInfoController(ApplicationDbContext context)
+        private const string authorizedRoleName = "Administrator";
+
+        public CustomerInfoController(ApplicationDbContext context, UserManager<ApplicationUser> userManager)
         {
-            _context = context;    
+            _context = context;
+            _userManager = userManager;
         }
 
         // GET: CustomerInfo
         public async Task<IActionResult> Index()
         {
+
+
             return View(await _context.Customers.ToListAsync());
         }
 
@@ -31,6 +38,14 @@ namespace RentalDataWarehouse.Controllers
             if (id == null)
             {
                 return NotFound();
+            }
+
+            if (!User.IsInRole(authorizedRoleName))
+            {
+                Guid currentUserCustomerId = (await _userManager.FindByNameAsync(User.Identity.Name)).CustomerId;
+
+                if (currentUserCustomerId != id)
+                    return Forbid();
             }
 
             var customer = await _context.Customers
@@ -87,12 +102,20 @@ namespace RentalDataWarehouse.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(Guid id, [Bind("Id,Name,EmailAddress")] Customer customer)
+        public async Task<IActionResult> Edit(Guid id, [Bind("Id,Name")] Customer customer)
         {
             if (id != customer.Id)
             {
                 return NotFound();
             }
+
+            if (!User.IsInRole(authorizedRoleName))
+            {
+                Guid currentUserCustomerId = (await _userManager.FindByNameAsync(User.Identity.Name)).CustomerId;
+
+                if (currentUserCustomerId != id)
+                    return new ForbidResult();
+            }            
 
             if (ModelState.IsValid)
             {
@@ -125,6 +148,9 @@ namespace RentalDataWarehouse.Controllers
                 return NotFound();
             }
 
+            if (!User.IsInRole(authorizedRoleName))
+                return new ForbidResult();
+
             var customer = await _context.Customers
                 .SingleOrDefaultAsync(m => m.Id == id);
             if (customer == null)
@@ -140,6 +166,9 @@ namespace RentalDataWarehouse.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(Guid id)
         {
+            if (!User.IsInRole(authorizedRoleName))
+                return new ForbidResult();
+
             var customer = await _context.Customers.SingleOrDefaultAsync(m => m.Id == id);
             _context.Customers.Remove(customer);
             await _context.SaveChangesAsync();
